Fixed circual dependency between api.js and storageService
