+++ eml/api/api.js
@@ -1,8 +1,6 @@
 import axios from 'axios';
 import { Buffer } from 'buffer';
 import { URL, CERTIFICATE_URL } from '@env';
-import { getStudentInfo } from './userApi';
-import { getUserInfo } from '../services/StorageService';
 
 const timeoutInMs = 1200;
 
@@ -185,78 +183,6 @@ export const unSubscribeToCourse = async (userId, courseId) => {
 };
 
 
-// Get certificates from student
-export const fetchCertificates = async (userId) => {
-	try {
-		if (userId == null) {
-			throw 'User ID is null';
-		}
-		const res = await axios.get(certificateUrl + '/api/student-certificates/student/' + userId);
-		return res.data;
-	} catch (e) {
-		if (e?.response?.data != null) {
-			throw e.response.data;
-		} else {
-			throw e;
-		}
-	}
-};
-
-
-/* Generates a certificate for a student.
- * @param {string} courseId - The ID of the course.
- * @param {string} studentId - The ID of the student.
- * @returns {Promise<Object>} - The response from the server.
- */
-export const generateCertificate = async (courseId, userId) => {
-	try {
-		// Fetch course data
-		const courseData = await getCourse(courseId);
-
-		// Fetch student data and user data concurrently
-		const [studentData, userData] = await Promise.all([
-			getStudentInfo(userId),
-			getUserInfo()
-		]);
-
-		// Ensure data is loaded
-		if (!courseData || !studentData || !userData) {
-			throw new Error('Course, student, or user data not loaded');
-		}
-
-		const object = {
-			courseName: courseData.title,
-			courseId: courseData._id,
-			studentId: studentData._id,
-			studentFirstName: userData.firstName,
-			studentLastName: userData.lastName,
-			courseCreator: courseData.creator,
-			estimatedCourseDuration: courseData.estimatedHours,
-			dateOfCompletion: new Date().toISOString().split('T')[0], // current date
-			courseCategory: courseData.category,
-		};
-
-		console.log(object);
-
-		// Call the endpoint to generate the certificate
-
-		const response = await axios.put(certificateUrl + '/api/student-certificates', {
-			courseName: courseData.title,
-			courseId: courseData._id,
-			studentId: studentData._id,
-			studentFirstName: userData.firstName,
-			studentLastName: userData.lastName,
-			courseCreator: courseData.creator,
-			estimatedCourseDuration: courseData.estimatedHours || 0,
-			dateOfCompletion: new Date().toISOString().split('T')[0], // current date
-			courseCategory: courseData.category,
-		});
-		return response.data;
-	} catch (error) {
-		console.error('Error generating certificate:', error.response?.data || error.message);
-		throw error;
-	}
-};
 
 export const giveFeedback = async (courseId, feedbackData) => {
 	const { rating, feedbackText, feedbackOptions } = feedbackData;
