+++ eml/api/userApi.js
@@ -12,7 +12,6 @@ const digitalOcean = 'http://207.154.213.68:8888';
 
 
 const url = URL;// Change this to your LOCAL IP address when testing.
-console.log('URL:', url);
 
 /**
  * This is the client that will be used to make requests to the backend.
